Move CTASignedInType/CTASignedOutType to module
