+++ front/app/modules/commercial/customizable_homepage_banner/admin/CTASettings/CTASignedInSettings.tsx
@@ -2,10 +2,9 @@ import React from 'react';
 import { SectionField } from 'components/admin/Section';
 import { FormattedMessage } from 'utils/cl-intl';
 import messages from '../messages';
-import {
-  CTASignedInType,
-  CustomizedButtonConfig,
-} from 'services/appConfiguration';
+import { CustomizedButtonConfig } from 'services/appConfiguration';
+import { CTASignedInType } from 'services/homepageSettings';
+
 import SettingRadioButtons from './SettingRadioButtons';
 import { CLErrors } from 'typings';
 import SettingsLabel from './SettingsLabel';
