+++ front/app/modules/commercial/customizable_homepage_banner/admin/CTASettings/CTASignedOutSettings.tsx
@@ -2,10 +2,8 @@ import React from 'react';
 import { SectionField } from 'components/admin/Section';
 import { FormattedMessage } from 'utils/cl-intl';
 import messages from '../messages';
-import {
-  CTASignedOutType,
-  CustomizedButtonConfig,
-} from 'services/appConfiguration';
+import { CustomizedButtonConfig } from 'services/appConfiguration';
+import { CTASignedOutType } from 'services/homepageSettings';
 import SettingRadioButtons from './SettingRadioButtons';
 import { CLErrors } from 'typings';
 import SettingsLabel from './SettingsLabel';
