+++ front/app/modules/commercial/customizable_homepage_banner/services/appConfiguration.ts
@@ -7,19 +7,6 @@ declare module 'services/appConfiguration' {
     two_row_layout: 'two_row_layout';
   }
 
-  interface CTASignedOutTypeMap {
-    sign_up_button: 'sign_up_button';
-    customized_button: 'customized_button';
-    no_button: 'no_button';
-  }
-  type CTASignedOutType = CTASignedOutTypeMap[keyof CTASignedOutTypeMap];
-
-  interface CTASignedInTypeMap {
-    customized_button: 'customized_button';
-    no_button: 'no_button';
-  }
-  type CTASignedInType = CTASignedInTypeMap[keyof CTASignedInTypeMap];
-
   interface CustomizedButtonConfig {
     text: Multiloc;
     url: string;
@@ -30,9 +17,7 @@ declare module 'services/appConfiguration' {
       allowed: boolean;
       enabled: boolean;
       layout: THomepageBannerLayout | null;
-      cta_signed_out_type: CTASignedOutType;
       cta_signed_out_customized_button?: CustomizedButtonConfig;
-      cta_signed_in_type: CTASignedInType;
       cta_signed_in_customized_button?: CustomizedButtonConfig;
     };
   }
