+++ front/app/services/homepageSettings.ts
@@ -4,19 +4,21 @@ import { Multiloc } from 'typings';
 // home_pages or homepages?
 const homepageSettingsEndpoint = `${API_PATH}/home_pages`;
 
+// type definitions
+
+// * THomepageSection *
 export interface IHomepageSectionMap {
   customizable_homepage_banner: 'customizable_homepage_banner';
   top_info_section_enabled: 'top_info_section_enabled';
   projects_enabled: 'projects_enabled';
   bottom_info_section_enabled: 'bottom_info_section_enabled';
 }
-
 export type THomepageSection = IHomepageSectionMap[keyof IHomepageSectionMap];
 
+// * THomepageBannerLayout *
 export interface THomepageBannerLayoutMap {
   full_with_banner_layout: 'full_width_banner_layout';
 }
-
 export type THomepageBannerLayout =
   THomepageBannerLayoutMap[keyof THomepageBannerLayoutMap];
 
@@ -32,20 +34,15 @@ export interface IHomepageSettings {
   customizable_homepage_banner: boolean;
   banner_layout;
   banner_signed_in_header_multiloc: Multiloc;
-  cta_signed_in_text_multiloc: Multiloc;
-  cta_signed_in_type;
-  cta_signed_in_url: string;
   banner_signed_out_header_multiloc: Multiloc;
   banner_signed_out_subheader_multiloc: Multiloc;
   banner_signed_out_header_overlay_color;
   banner_signed_out_header_overlay_opacity;
-  cta_signed_out_text_multiloc: Multiloc;
-  cta_signed_out_type;
-  cta_signed_out_url: string;
   header_bg;
   pinned_admin_publication_ids: string[];
 }
 
+// streams
 export function homepageSettingsStream() {
   return streams.get<IHomepageSettings>({
     apiEndpoint: homepageSettingsEndpoint,
